% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_event_performance.R
\name{eval_event_plot_tsline_cum_ret}
\alias{eval_event_plot_tsline_cum_ret}
\title{Plot event cumulative-return bands}
\usage{
eval_event_plot_tsline_cum_ret(res, threshold = 0.02)
}
\arguments{
\item{res}{result from \code{eval_event_performance()}.}

\item{threshold}{numeric; dashed reference lines at \eqn{\pm}threshold.}
}
\value{
ggplot object.
}
\description{
Plot event cumulative-return bands
}
