% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-yieldcurve.R
\name{get_yield_data_DT}
\alias{get_yield_data_DT}
\title{Get yield data for given dates}
\usage{
get_yield_data_DT(yield_dt_list, yield_dates)
}
\arguments{
\item{yield_dt_list}{A list of data.tables, each containing a yield series with columns \code{date} and \code{value}.}

\item{yield_dates}{A vector of dates for which yields should be retrieved.}
}
\value{
A data.table containing yield data across maturities and dates.
}
\description{
Get yield data for given dates
}
\examples{
series <- lapply(seq_len(11), function(i) {
  data.table::data.table(date = as.Date(c("2025-01-01", "2025-02-01")),
                         value = c(4 + i / 20, 3.8 + i / 20))
})
curves <- get_yield_data_DT(series, list(
  Now = as.Date("2025-02-01"), Previous = as.Date("2025-01-01")
))
gen_yield_curve_plot(curves, selected_windows = c("Now", "Previous"))
gen_yield_curve_plot_grid(curves, selected_windows = c("Now", "Previous"))
}
