% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_strat_simple_backtesting.R
\name{eval_strat_plot_tsline_eq}
\alias{eval_strat_plot_tsline_eq}
\title{Plot equity curve (time series)}
\usage{
eval_strat_plot_tsline_eq(bt_res)
}
\arguments{
\item{bt_res}{single-row data.table from \code{eval_strat_performance()}.}
}
\value{
ggplot object.
}
\description{
Plot equity curve (time series)
}
