% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plots.R
\name{gen_yield_curve_plot}
\alias{gen_yield_curve_plot}
\title{Generate U.S. Treasury yield curve plot}
\usage{
gen_yield_curve_plot(
  DT,
  selected_windows = c("Now", "One week ago", "One month ago")
)
}
\arguments{
\item{DT}{A data.table returned from \code{get_yield_data_DT()}.}

\item{selected_windows}{Character vector specifying which yield curves to display (e.g., \code{"Now"}, \code{"One week ago"}).}
}
\value{
A ggplot2 object visualizing selected yield curves.
}
\description{
Generate U.S. Treasury yield curve plot
}
