% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plots.R
\name{gen_yield_curve_plot_grid}
\alias{gen_yield_curve_plot_grid}
\title{Generate yield curve comparison grid}
\usage{
gen_yield_curve_plot_grid(
  DT,
  selected_windows = c("Now", "One week ago", "One month ago", "Six months ago",
    "One year ago")
)
}
\arguments{
\item{DT}{A data.table returned from \code{get_yield_data_DT()}.}

\item{selected_windows}{Character vector specifying yield curve snapshots for comparison.}
}
\value{
A faceted ggplot2 object comparing yield curves across multiple time windows.
}
\description{
Generate yield curve comparison grid
}
