% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plots.R
\name{get_yield_data_DT}
\alias{get_yield_data_DT}
\title{Get yield data for given dates}
\usage{
get_yield_data_DT(yield_dt_list, yield_dates)
}
\arguments{
\item{yield_dt_list}{A list of data.tables, each containing a yield series with columns \code{date} and \code{value}.}

\item{yield_dates}{A vector of dates for which yields should be retrieved.}
}
\value{
A data.table containing yield data across maturities and dates.
}
\description{
Get yield data for given dates
}
