% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-forecast.R
\name{prep_fill_forward}
\alias{prep_fill_forward}
\title{Carry the latest observed value forward}
\usage{
prep_fill_forward(x)
}
\arguments{
\item{x}{Vector containing values and missing observations.}
}
\value{
A vector with missing values filled from the latest prior observation.
}
\description{
Carry the latest observed value forward
}
