% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-style.R
\name{viz_annotate_last_value}
\alias{viz_annotate_last_value}
\title{Add a direct label for the last observation in a series}
\usage{
viz_annotate_last_value(
  plot,
  data,
  x,
  y,
  label = NULL,
  prefix = "",
  digits = 2,
  color = NULL,
  style = NULL,
  context = NULL
)
}
\arguments{
\item{plot}{A `ggplot` object.}

\item{data}{A data.frame or data.table containing the plotted series.}

\item{x}{Name of the x column.}

\item{y}{Name of the y column.}

\item{label}{Optional name of a label column. If `NULL`, the formatted `y` value is used.}

\item{prefix}{Optional prefix prepended to the label text.}

\item{digits}{Number of digits used when formatting numeric labels.}

\item{color}{Optional label color.}

\item{style}{Named style family or `NULL` to use the package default.}

\item{context}{Output context or `NULL` to use the package default.}
}
\value{
A `ggplot` object with a last-value annotation.
}
\description{
Add a direct label for the last observation in a series
}
