% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-style.R
\name{viz_palette_get}
\alias{viz_palette_get}
\title{Get visualization palette values}
\usage{
viz_palette_get(
  style = NULL,
  context = NULL,
  palette = c("discrete", "sequential")
)
}
\arguments{
\item{style}{Named style family or `NULL` to use the package default.}

\item{context}{Output context or `NULL` to use the package default.}

\item{palette}{One of `"discrete"` or `"sequential"`.}
}
\value{
Character vector of colors.
}
\description{
Get visualization palette values
}
